SOLR-15583 Backport bugs from error-prone 2.8.1

+++ solr/solrj/src/java/org/apache/solr/client/solrj/io/eval/RawValueEvaluator.java
@@ -54,7 +54,7 @@ public class RawValueEvaluator extends SourceEvaluator {
   
   private void init(Object value){
     if(value instanceof Integer){
-      this.value = (Long)value;
+      this.value = ((Integer)value).longValue();
     }
     else if(value instanceof Float){
       this.value = ((Float)value).doubleValue();
